requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiohttp-retry==2.4.6
aiosignal==1.2.0
alembic==1.4.1
altair==4.1.0
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==21.1.0
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.1
atpublic==2.3
attrs==21.2.0
backcall==0.2.0
base58==2.1.1
bleach==4.1.0
blinker==1.4
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==7.1.2
cloudpickle==2.0.0
colorama==0.4.4
commonmark==0.9.1
configobj==5.0.6
cycler==0.11.0
databricks-cli==0.16.2
DateTime==4.3
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
dictdiffer==0.9.0
diskcache==5.3.0
distro==1.6.0
docker==5.0.3
dpath==2.0.5
dulwich==0.20.26
dvc==2.8.3
entrypoints==0.3
Flask==2.0.2
flatbuffers==2.0
flatten-dict==0.4.2
flufl.lock==6.0
fonttools==4.28.3
frozenlist==1.2.0
fsspec==2021.11.1
ftfy==6.0.3
funcy==1.16
future==0.18.2
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.24
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grandalf==0.6
greenlet==1.1.2
grpcio==1.42.0
gunicorn==20.1.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.8.2
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.2.1
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
lightgbm==3.3.1
llvmlite==0.37.0
mailchecker==4.1.3
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
missingno==0.5.0
mistune==0.8.4
mlflow==1.22.0
multidict==5.2.0
nanotime==0.5.2
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
notebook==6.4.6
numba==0.54.1
numpy==1.20.3
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
phonenumbers==8.12.38
pickleshare==0.7.5
Pillow==8.4.0
ply==3.11
prometheus-client==0.12.0
prometheus-flask-exporter==0.18.6
prompt-toolkit==3.0.23
protobuf==3.19.1
psutil==5.8.0
ptyprocess==0.7.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.7.1
pydot==1.4.2
pygit2==1.7.1
Pygments==2.10.0
pygtrie==2.4.2
Pympler==0.9
pyparsing==3.0.6
pyrsistent==0.18.0
python-benedict==0.24.3
python-dateutil==2.8.2
python-editor==1.0.4
python-fsutil==0.5.0
python-slugify==5.0.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.1
QtPy==1.11.3
querystring-parser==1.2.4
requests==2.26.0
requests-oauthlib==1.3.0
rich==10.15.2
rsa==4.8
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
scikit-learn==1.0.1
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
setuptools-scm==6.3.2
shap==0.40.0
shortuuid==1.0.8
shtab==1.5.2
six==1.16.0
sklearn==0.0
slicer==0.0.7
smmap==5.0.0
SQLAlchemy==1.4.27
sqlparse==0.4.2
streamlit==1.2.0
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
threadpoolctl==3.0.0
toml==0.10.2
tomli==1.2.2
toolz==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
validators==0.18.2
voluptuous==0.12.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
wrapt==1.13.3
xgboost==1.5.1
xmltodict==0.12.0
yarl==1.7.2
zc.lockfile==2.0
zipp==3.6.0
zope.interface==5.4.0
